requirements.txt
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
credentials==1.1
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.6
docopt==0.6.2
filelock==3.8.0
firebase-admin==6.0.1
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
httplib2==0.20.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
platformdirs==2.5.2
ply==3.11
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.56
PyJWT==2.5.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
types-cryptography==3.3.23.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.9.0
pillow==10.1.0
